requirements.txt
apache-airflow==2.10.2
duckdb
pyarrow
pandas